+++ apps/mobile/src/hooks/useReturnToOnAuthPath.js
@@ -1,5 +1,6 @@
 import { useEffect } from 'react'
 import { isEmpty } from 'lodash/fp'
+import * as Sentry from '@sentry/react-native'
 import useLinkingStore from 'navigation/linking/store'
 import { openURL } from 'hooks/useOpenURL'
 
@@ -10,7 +11,7 @@ export default function useReturnToOnAuthPath (loading = false) {
     (async function () {
       if (!loading && !isEmpty(returnToOnAuthPath)) {
         setReturnToOnAuthPath()
-        await openURL(returnToOnAuthPath)
+        await openURL(returnToOnAuthPath).catch(err => Sentry.captureException(err, { extra: { returnToOnAuthPath } }))
       }
     })()
   }, [